[EngSys] Add missing `ts-node` dev dependencies
Some projects are missing `ts-node` dev dependencies even though they are
depending it to run tests in .ts files. They work today probably because PNPM
store dependencies in a way that allows `ts-node` to be resolved even it's not
in the devDependencies list.  However, it may not work in all scenarios. For
example, after migrating a packge to ESM.

This PR adds the missing `ts-node` dev dependencie and also make the version
consistent (^10.0.0)