Re merging Hierarchical Partition feature to main branch (#26131)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
- Re merging Hierarchical Partition feature to main.
Added Hierarchical Partition Key Support

1. Refactored Definition of PartitionKey.
- Old definition - type PartitionKey = PartitionKeyDefinition | string |
number | unknown;
- New definition - type PartitionKey = PrimitivePartitionKeyValue |
PrimitivePartitionKeyValue[];
   - The main change here is
- Separating out PartitionKeyDefinition [May seem like a breaking
change. But the actual functions which need PartitionKeyDefinition were
always taking PartitionKeyDefinition as an input i.e.
ContainerDefinition, the existing definition seems confusing.]
- Removal of unknown from definition (along with adding boolean).
Defining it as PartitionKeyDefinition | string | number | unknown was
effectively equivalent to unknown, since or of anything with unknown is
unknown. Which essentially means we had no structure upon PartitionKey.
This seems like a huge bug.
2. Introduced PartitionKeyInternal to create a boundary beyond with
PartitionKey would be sanitized. The idea is to use ClientContext as
that boundary.
3. Changed bulk/batch api Operation/OperationInput DTOs partitionKey
field to PartitionKey type.
4. Clearly defined order for choosing partitionKey for an operation
during bulk api. Now if the user has provided partitionKey in
OperationInput it will be used, if not it will be derived from resource
body (for create and upsert).
5. Added logic for MultiHash calculation.
6. Changed FeedOption.partitionKey filed to PartitionKey type.
7. Unrelated changes that are part of the PR (due to unrelated bugs)
- Adding check based on the format of requestCharge present in
FeedResponse. Removed bounds for RU in cross partition test case, due to
it being flaky.

### Describe the problem that is addressed by this PR


### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: FAREAST\vikassingh <vikassingh@microsoft.com>