[@azure/eventgrid] Fix the Push Issue with the PublishCloudEvents API (#26307)
### Packages impacted by this PR
@azure/eventgrid

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR
In the EventGrid V2 Beta Version, we have an API called
`publishCloudEvents`. In this API, each of the event (from parameters)
must be converted and pushed to the `eventsWireModel` array. Now, while
pushing it, a wrong method is called on the array. Instead of calling
`push` method, `concat` method is called which causes the array to be
empty and causes issue. This PR changes the method to `push`.

Integration Test Cases have been added for each API in EventGrid Client.
The associated recordings have been added to the PR.

**Note: Since this is a Bug Fix release, I have updated the Version from
`4.13.0-beta.1` to `4.13.0-beta.2`.**

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
The design is simple and straightforward. No special concern.

### Are there test cases added in this PR? _(If not, why?)_
Yes. The test cases have been added for each API of the EventGrid
client.

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
N/A

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

@ellismg @xirzec @jeremymeng  Please review and approve the PR.

@lmazuel FYI....