[@azure/confidential-ledger-rest] Migrate to latest test recorder and externalize recordings (#26305)
### Packages impacted by this PR

`@azure-rest/confidential-ledger`
`@azure-tools/test-recorder`

### Issues associated with this PR

Fixes #20495

### Describe the problem that is addressed by this PR

Migrate to the latest test recorder in order to use the test proxy and
externalize recordings. I also had to regenerate the RLC in order to fix
an issue with `isUnexpected` not working properly for the user endpoint.

Getting the test resources to deploy correctly also took some doing
since it uses the `oid` and not the normal client id of the test app.

The tests themselves were a bit busted in places; I shored them up as
best I could given the odd behavior of the service.

Lastly, the test recorder didn't have support for any custom transport
options, so I had to add one in order to pass the `TLSValidationCert` in
order for it to work properly with confidential ledger.