[quantum-jobs] Migrate to test proxy and delete old recordings (#26316)
### Packages impacted by this PR

`@azure/quantum-jobs`

### Describe the problem that is addressed by this PR

Due to the expense and difficulty of running tests against Quantum Jobs,
I have not been able to validate these changes, but given this package
has not been updated since its initial beta release in 2/5/2021, it's
unlikely going to ship again soon. If and when it does ship we can
ensure the tests are properly updated and recorded.