[ai-metrics-advisor] Migrate to test proxy and temporarily disable tests (#26330)
### Packages impacted by this PR

`@azure/ai-metrics-advisor`

### Issues associated with this PR

Closes #18001

### Describe the problem that is addressed by this PR

This PR migrates `ai-metrics-advisor` to the latest version of the test
recorder so we can fully deprecate the 1.0 version. Due to the tests not
currently passing in live (as they require static resources that are
difficult to maintain), the tests are disabled for now until we can
figure out how to address this issue prior to a future release.