Ericasp/add number lookup to communication phone numbers (#26254)
communication-phone-numbers

https://github.com/Azure/azure-sdk/issues/6321

Adding the SearchOperatorInformation method to the SDK to enable access
to the new Number Lookup feature in private preview

there are more than one possible design, why was the one in this PR
chosen?
We considered adding the new Number Lookup feature to its own SDK but
decided to put it with the existing Phone Numbers SDK for clarity since
it's conceptually related to phone numbers and we didn't want to create
a new client for one endpoint.

Yes

https://github.com/Azure/azure-sdk-for-net/pull/35574

request PRs)_
n/a

- [ ] Added impacted package name to the issue description
- [n/a] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [n/a] Added a changelog (if necessary)