Add missing service label entries to CODEOWNERS (#26500)
The FabricBot rule for @ mentioning people based upon the service label
when the Service Attention label was added should have been generated
from the CODEOWNERS file for each repository. At some point and time
this stopped happening and changes were made directly to FabricBot. What
I'm doing is getting the last FabricBot.json file for this repository,
that was removed as part of the [PR where github-event-processor was
enabled](https://github.com/Azure/azure-sdk-for-js/pull/25511) and
adding any entries that were in the FabricBot rule that are missing from
the CODEOWNERS file.