[iot-modelsrepository][test] fix browser test failure (#26638)
Browser test runs for iot have been failing for a while. It was working
likely due to some other packages installed test-recorder and karma was
able to resolve it. However, it stopped working since the end of June.

This PR adds the missing dev dependency of `@azure-tools/test-recorder`
and move karma.conf.js to be consistent with other packages.