Thread metadata (#26797)
### Packages impacted by this PR

@azure/communication-chat

### Issues associated with this PR
https://github.com/Azure/azure-sdk/issues/6245

With these API changes, we are giving developers the ability to store
additional contextual information about chat threads and chat
participants. This metadata can give the flexibility to our customers to
implement logic specific to their use cases and systems. For example,
customers could link different related chat threads to each other and
specify the thread types according to their own scenarios.

This metadata could also be used to pass in contextual information that
other chat participants - call center agents, bots - can use to
communicate on the chat threads (e.g. locale and timezone).

Special considerations: The Participant API doesn't have a PATCH yet. It
will be added in a future API version. This is an existing problem for
updating Participant display names.

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)