[@azure/monitor-query] Code Changes to add scope to Metrics Query Client and modify MetricsBatchOptionalParams (#26868)
### Packages impacted by this PR
@azure/monitor-query

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR
1. In this release, In the `MetricsBatchOptionalParams`object, the names
of the properties `endtime`, `orderby`, `starttime` are changed to
`endTime`, `orderBy`, `startTime` respectively.
2. The scope value is not set correctly in the previous release. In this
release, I am setting the scope value correctly.
3. The previous release did not have integration tests. With this PR, I
have added the integration tests.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
There are no special design changes in this PR, other than setting the
credential scope.

### Are there test cases added in this PR? _(If not, why?)_

One point to note is that the new tests are marked as skip. Because, the
recordings will have subscription id in them which we do not want to
expose. So, for CI, I have marked them as Skip. I am working on possible
solutions to bypass this issue in the future. But, the tests have been
run locally and the snapshots are provided below.


![image](https://github.com/Azure/azure-sdk-for-js/assets/602456/43e9149b-a882-4707-b43b-68b3f4df7211)


![image](https://github.com/Azure/azure-sdk-for-js/assets/602456/f89e32a2-2af9-45b6-a90e-1e1ddc504166)

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
N/A

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

---------

Co-authored-by: Scott Addie <10702007+scottaddie@users.noreply.github.com>