[core-http] Port changes from PR #26897 Defer Error construction
`@azure/core-http`

Partner report of memory leak in test bench.

Eager error construction on module load can cause trouble when the Error
constructor is hooked/cached by development tooling. This PR changes the
pattern of caching the Error object to a more typical new-on-throw
pattern. This should also improve the stack trace to be more accurate on
the thrown Error.