[core-http] Port changes from PR #26897 Defer Error construction (#26902)
### Packages impacted by this PR

`@azure/core-http`

### Issues associated with this PR

Partner report of memory leak in test bench.

### Describe the problem that is addressed by this PR

Eager error construction on module load can cause trouble when the Error
constructor is hooked/cached by development tooling. This PR changes the
pattern of caching the Error object to a more typical new-on-throw
pattern. This should also improve the stack trace to be more accurate on
the thrown Error.