[OpenAI] Add Whisper (#27109)
### Packages impacted by this PR
@azure/openai

### Issues associated with this PR
None for whisper but has a rudimentary fix for
https://github.com/Azure/azure-sdk-for-js/issues/26953

### Describe the problem that is addressed by this PR
Adds support for speech to text capabilities. See the changelog entry
and the samples for more details about the addition.

Few notes:
- Bring Your Own Data tests are skipped because the new version
deployment doesn't support it yet, hopefully the support should be there
soon
- @azure/core-rest-pipeline's `formDataPolicy` doesn't support file
uploads. I added a custom version of the policy in openai that supports
file uploads and uses an actively maintained 3rd party library.
- adds a fix for https://github.com/Azure/azure-sdk-for-js/issues/26953
that doesn't rely on core changes (see the changes in
`src/api/getSSE.ts` and `src/api/getSSE.browser.ts` files. A better fix
is in https://github.com/Azure/azure-sdk-for-js/pull/27000 but that is
still being reviewed.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
N/A

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)

---------

Co-authored-by: Minh-Anh Phan <111523473+minhanh-phan@users.noreply.github.com>