[recorder] Fix env var issue where one variable a substring of another (#27187)
### Packages impacted by this PR

- `@azure-tools/test-recorder`

### Issues associated with this PR

- Fixes #11379 (an issue from 2020!)

### Describe the problem that is addressed by this PR

Fixes a long-running issue recently rediscovered by @minhanh-phan. If we
have 2 environment variables in our `.env`, where one is a substring of
another, e.g.:

```sh
VAR_1=TestTestTest
VAR_2=Test
```

we can run into an issue where the shorter environment variable gets
replaced first in the recordng, causing the first variable to not be
sanitized properly. This causes issues in playback.

This PR fixes the issue by:
- Ordering the sanitizers applied by `envSetupForPlayback` by the length
of the environment variable value, in descending order, and
- Adding the sanitizers in a deterministic order instead of applying
them in parallel and using `Promise.all`.