[OpenAI] Bundle using rollup v3 (#27258)
### Packages impacted by this PR
@azure/openai

### Issues associated with this PR
Fixes https://github.com/Azure/azure-sdk-for-js/issues/27227
Fixes https://github.com/Azure/azure-sdk-for-js/issues/27226

### Describe the problem that is addressed by this PR
`@azure/openai` depends on v4 of `formdata-node` and v1.7.2 of
`form-data-encoder`. However, those are outdated versions and we should
depend on recent versions of those libraries instead but the recent
versions are ESM-only. I couldn't find a way to bundle ESM code into cjs
using rollup v2, our current rollup version. With rollup v3, you can
bundle ESM-only code if you import it with dynamic imports and this PR
does exactly that.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
N/A

### Are there test cases added in this PR? _(If not, why?)_
N/A

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)