[ServiceBus] throw error earlier when timing out
Related: issue #9775, #27010

This PR fixed the following timing issue when trying to establish link for next
session but there are none available.

what's happening is

- SDK is trying to create a link for next avaiable session (`options.source.filter: { com.microsoft:session-filter': undefined }`)
- Because no sessions are available, after one minute, a link is established,
  but with `"source":null,"target":null`. The call to create link resolves.
- However, the service immediately use that link to ask SDK to close the link, with an error of `"com.microsoft:timeout"`/`The operation did not complete within the allotted timeout of 00:00:59.9130000.`
- SDK respects the request and start acquiring a lock to close the link
- Meanwhile, SDK gets a link, nice! It will continue as the link has been created
- SDK gets lock to close link and sets `this._link` to `undefined` and start closing receiver and amqp session
- SDK null-check the link and throws INTERNAL ERROR

Instead of returning the link, This PR checks whether we got a session id back.
If not and there's an error from service, we rethrow the error.  Some of the
existing error handling is also moved before returning the link, instead of
afterwards.