Cherry pick from main PR #27302 from main branch
[engsys] update platform matrix for node 14 deprecation (#27302)

Update the CI matrix so we are no longer testing on Node 14 as it has
reached the end of our EOL + 6 months policy.

For coverage and browser legs I bumped it to use the current LTS (18)
rather than 16 which is EOL, since most customers are not expected to be
using an out of support runtime.