[Event Hubs] Use dev-tool's bundle (#27305)
Successful live runs:
- [Schema
Registry](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=3142121&view=results)
- [Event
Hubs](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=3143866&view=results)

### Packages impacted by this PR
@azure/event-hubs

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/26665

### Describe the problem that is addressed by this PR
Move from rollup v2 to v3.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
Node polyfilling is not doing the right thing in Event Hubs for some
reason and I had to disable it. In particular, the buffer API was
causing bad generated tokens that are used for SAS authentication with
the service. My approach is to continue to rely on the inject plugin to
inject imports to the polyfill buffer dependency. However, since we
disabled node polyfilling for event hubs, we have to do manual
polyfilling for all libraries that depend on event hubs which are
basically the schema registry serializers. This PR manually polyfills
streams used by avsc and suppresses rollup warnings for missing other
node builtins since they're not needed in our scenarios.

The advantage of this work is that we no longer need custom rollup
configurations for Event Hubs or Schema Registry.

### Are there test cases added in this PR? _(If not, why?)_
N/A

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)