[core-util] Fix issue with Deno showing up as isNode at runtime. (#27459)
### Packages impacted by this PR

`@azure/core-util`

### Issues associated with this PR

Fixes #27077

### Describe the problem that is addressed by this PR

Deno implemented `process.versions.node`, which is how we detect that we
are running inside a Node environment. Since Deno actually is much
closer to a browser environment, this is *not* what was expected at
runtime and resulted in things like loading the browser version of
pipeline policies that throw due to not being implemented/relevant in
browsers.

I kept this really simple by making `isNode` check that `isDeno` is
false. Since `isNode` was previously false inside Deno environments, I
don't think this will break anything.

### Provide a list of related PRs _(if any)_

https://github.com/Azure/azure-sdk-for-js/pull/27090