[core-util] Add support for hex strings to `uint8ArrayToString` (#27441)
### Packages impacted by this PR

- `@azure/core-util`

### Describe the problem that is addressed by this PR

Key Vault needs a util to convert byte arrays to a hex string for an
upcoming convenience property. This adds that support to the existing
`core-util` `uint8ArrayToString` helper.

In this PR I also simplified the logic for `uint8ArrayToString` in node,
and also removed the now-redundant Base64 and hex helpers that aren't
exported but are used elsewhere in `core-util`.

This PR also fixes a bug with the browser Base64 `uint8ArrayToString`
implementation where the conversion would throw upon encountering bytes
which are not valid ISO-8859-1 (Latin1) code points.