[ServiceBus] Throw a retryable error when sending message and link is detached
It's possible that after we create the sender link, but before using it to send
message, the link gets closed for example due to bad network condition. Our
current code would throw a type error when accessing `send()` method of the
undefined `link`. The error is not retryable.

This PR checks for valid link again before calling `link.send()` and throw an
new retryable client error.