[engsys] Update CI to use node 18 as a minimum runtime version (#27553)
### Describe the problem that is addressed by this PR

This PR migrates our CI away from testing Node16. 

Due to Node16 [depending on an insecure version of
OpenSSL](https://nodejs.org/en/blog/announcements/nodejs16-eol) we
believe that customers should migrate without delay to a supported LTS
version of the Node runtime. As such we do not intend to fix bugs that
only reproduce in Node 16 runtimes, obviating the need to proactively
run in this environment.