[@azure/eventgrid] Fixing the Readme issue (#27489)
### Packages impacted by this PR

@azure/eventgrid

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/26367

### Describe the problem that is addressed by this PR
The sample provided in the Readme is both complex and not working. I
have updated the sample with a simple scenario which highlights the API
that is explained. I have also verified that this new sample works fine.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
No special consideration for design.

### Are there test cases added in this PR? _(If not, why?)_
No. This is just updating the Readme file. Also, the sample is tested.

### Provide a list of related PRs _(if any)_
NA

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
NA

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

@xirzec Please review and approve the PR