[OpenAI] Use core's formdata policy (#27558)
### Packages impacted by this PR
@azure/openai

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR

It moves OpenAI off unnecessary 3rd party dependencies now that our core
can handle form data content types more effectively.

~This PR moves the improved form data policy in OpenAI to core to
decrease the custom layer in OpenAI and to strengthen our core
libraries. The policy is improved as follows:~
- ~ditches [`form-data`](https://www.npmjs.com/package/form-data) which
has not been in active development for 3 years. It is also not compliant
with the [form data
spec](https://developer.mozilla.org/en-US/docs/Web/API/FormData).~
- ~it supports custom boundaries, so a boundary can be fixed to enable
playing tests back.~

~Note that the boundary format changed from being a random number to a
random string: For example: "--form-data-boundary-lj1n3jyl3fk2u3a8--".
This is fine because the spec specifies the boundary to be of no more
than 70 bytes long of 7-bit US-ASCII characters.~

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
~Alternatively, we can rely on the fetch client but there is still a
question about its performance. This is tracked in
https://github.com/Azure/azure-sdk-for-js/issues/27427.~

### Are there test cases added in this PR? _(If not, why?)_
N/A

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)