[ServiceBus] await error handler in session receiver (#27716)
Currently in session receiver, we fire-and-forget user's error handler
instead of awaiting it. This is causing problems when customer want to
modify state in message handler, and want to roll back in case of error
as currently we would carry on to abandon the message when error
happens, then add credit. The same message could arrive and processed,
while the error handling is still going on.

This PR changes it to wait for user error handler before proceeding.
This behavior of awaiting user error handler is in line with our
streaming receiver and other languages.