[@Azure/eventgrid] Adding new events for EG Version 5.0.0 (#27679)
### Packages impacted by this PR
@azure/eventgrid

**API View To Approve** 

https://apiview.dev/Assemblies/Review/30d6488f3dd647a79d175e408d5d5303?diffRevisionId=20bf158c9f2341eba08c706d55a0aa4c&diffOnly=False&revisionId=f5dd29153bf946c0aa1d47886fcd5eaf&doc=true

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

The EventGrid Service Team has added 2 new events. The SDK must be
regenerated to include the code changes related to these 2 new events.
The SDK minor version has been updated.

- `Microsoft.ResourceNotifications.Resources.CreatedOrUpdated`
- `Microsoft.ResourceNotifications.Resources.Deleted`


### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

There are no specific/complex design scenarios for this task. It is a
straightforward regenerate and some standard changes to the custom layer
of the code.

One of the important issues to note down is that this release does have
a breaking change. The data type of `tags` property has been changed
from `string` to object. So, this is a breaking change. Accordingly, the
major version has been updated.

### Are there test cases added in this PR? _(If not, why?)_

No. This item is standard and we need not add test cases for every new
events. The existing cases would be sufficient.

### Provide a list of related PRs _(if any)_

- https://github.com/Azure/azure-sdk-for-js/pull/27384 (This is the PR
that adds similar events to the SDK in the 4.15.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/26939 (This is the PR
that adds similar events to the SDK in the 4.14.0 release)
- https://github.com/Azure/azure-sdk-for-js/pull/26020 (This is the PR
that adds similar events to the SDK in the 4.13.0 release)

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

`autorest --typescript swagger\README.md`

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

@xirzec Please review and approve the PR and API View. Thanks