[@azure/eventgrid] - Azure EventGrid Beta Release V4.13.0-beta.3 (#27666)
### Packages impacted by this PR

@azure/eventgrid

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

1. The EventGrid Service team has made changes to their [CADL
specification](https://github.com/Azure/azure-rest-api-specs/tree/main/specification/eventgrid/Azure.Messaging.EventGrid).
This PR regenerates the SDK using the latest CADL specification.
2. Important changes in this PR:
     i. A new API `renewCloudEventLocks` has been added.
ii. In the ` publishCloudEvent` API, a new option `binaryMode` has been
added. This enables the users to send in the binary encoded event.
iii. There are some other minor changes and some breaking changes also
in this PR.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

One of the design decisions is around the `binaryMode`. Currently, the
API accepts only binary encoded data for this API. If there is a
requirement in the future that the user prefers that the encoding be
done by the SDK, we could consider it in future,

### Are there test cases added in this PR? _(If not, why?)_

Yes. New test cases have been added to test the `binaryMode` and the new
`renewCloudEventLocks` API.

### Provide a list of related PRs _(if any)_
1. https://github.com/Azure/azure-sdk-for-js/pull/26307
2. https://github.com/Azure/azure-sdk-for-js/pull/25906

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

@xirzec Please review and approve the PR & API Views.

## API View to approve

1. [Link
1](https://apiview.dev/Assemblies/Review/0ead6d0c15c64cd7a007a11f514cfc4c/b94b6c83619f4dcfb04e476365bf7e19?diffRevisionId=fb754e995af24441901461c1c6ada906&doc=False&diffOnly=True)
(This compares this Beta 3 version with Beta 2 version which gives you a
clear view of the changes)
2. [Link
2](https://apiview.dev/Assemblies/Review/b6a1a4040a1f46528f095df8589bfa3e/80e3111615cc47f49b25a4f16d9e1322?diffRevisionId=20bf158c9f2341eba08c706d55a0aa4c&doc=False&diffOnly=True)
(This is generated automatically by the PR Bot wgich compares the Beta 3
version with latest GA which is not correct. But, in order to release
the SDK, we need the approval for the same. So, kindly review the Link 1
and approve both). Thanks