[core-client-rest] keep request body as bytes for octet-stream
Currently we convert request body to string by default. This isn't always
desirable and could be problematic. It breaks Image Analysis scenario where
request body needs to be raw bytes.

This PR keeps the raw bytes as the request body when the content type is
`applicatoni/octect-stream`. This is a short-term fix to unblock Image Analysis
SDK while we investigate for better and long term solutions.