[engsys] Avoid running eslint plugin unit tests on every analyze (#27938)
### Describe the problem that is addressed by this PR

I noticed that we were running the unit tests for
`@azure/eslint-plugin-azure-sdk` as part of the 'run-eslint' step that
gets kicked off as part of `analyze` for every SDK package. This meant
we spent 1 minute 30 seconds running unit tests each and every time we
ran CI for any package.

My (perhaps incomplete) understanding is that we added running the tests
so that the ci.yml for `@azure/eslint-plugin-azure-sdk` could run the
tests as part of CI, but I don't think we intended to have them run all
the time for everyone.

This change creates a new `test-eslint` template that can be used by
`ci.yml` and removes running the unit tests from the normal analyze
step.